+++ src/app.module.ts
@@ -5,7 +5,7 @@ import { AuthModule } from './auth/auth.module';
 import { UserModule } from './user/user.module';
 import { RecipeModule } from './recipe/recipe.module';
 import { PrismaModule } from './prisma/prisma.module';
-import { RedisCacheModule } from './cache/redisCache.module';
+import { RedisCacheModule } from './cache/redis-cache.module';
 
 @Module({
   imports: [
